requirement.txt
asgiref==3.3.4
dj-config-url==0.1.1
dj-database-url==0.5.0
Django==3.2
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.7.1
gunicorn==20.1.0
Markdown==3.3.4
numpy==1.20.2
opencv-python==4.5.1.48
Pillow==8.2.0
PyJWT==1.7.1
PyMySQL==1.0.2
pypng==0.0.20
pytz==2021.1
pyzbar==0.1.8
sqlparse==0.4.1
